requirements.txt
scrapy==2.7.0
mysql-connector-python
python-dotenv
